+++ src/Bozboz/Admin/Services/Validators/EmptyValidator.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Bozboz\Admin\Services\Validators;
+
+use Bozboz\Admin\Services\Validators\Validator;
+
+class EmptyValidator extends Validator
+{
+	protected function passes($attributes, $rules)
+	{
+		return true;
+	}
+}