+++ src/Base/Composers/Nav.php
@@ -1,8 +1,8 @@
-<?php namespace Bozboz\Admin\Composers;
+<?php namespace Bozboz\Admin\Base\Composers;
 
 use Event;
 use Auth;
-use Bozboz\Admin\Components\Menu;
+use Bozboz\Admin\Base\Components\Menu;
 
 class Nav
 {