+++ src/Media/MediaAdminDecorator.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Bozboz\Admin\Decorators;
+namespace Bozboz\Admin\Media;
 
+use Bozboz\Admin\Base\ModelAdminDecorator;
 use Bozboz\Admin\Fields\BelongsToManyField;
 use Bozboz\Admin\Fields\ListField;
+use Bozboz\Admin\Fields\MediaField;
 use Bozboz\Admin\Fields\SelectField;
 use Bozboz\Admin\Fields\TextField;
 use Bozboz\Admin\Reports\Filters\ArrayListingFilter;
 use Bozboz\Admin\Reports\Filters\MultiOptionListingFilter;
 use Bozboz\Admin\Reports\Filters\SearchListingFilter;
-use Bozboz\Admin\Fields\MediaField;
-use Bozboz\Admin\Models\Media;
 
 class MediaAdminDecorator extends ModelAdminDecorator
 {