+++ src/Media/MediableTrait.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Bozboz\Admin\Media;
+
+trait MediableTrait
+{
+	public function media()
+	{
+		return Media::forModel($this);
+	}
+}