+++ src/Models/Exceptions/UploadException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Bozboz\Admin\Exceptions;
-
-class UploadException extends \Exception
-{
-
-}