+++ src/Services/Uploader.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Bozboz\Admin\Models;
+namespace Bozboz\Admin\Services;
 
 use Bozboz\Admin\Exceptions\UploadException;
-use Bozboz\Admin\Models\Media;
+use Bozboz\Admin\Media\Media;
 use Illuminate\Support\Str;
 use Symfony\Component\HttpFoundation\File\UploadedFile;