+++ src/Bozboz/Admin/Reports/NestedReport.php
@@ -14,9 +14,8 @@ class NestedReport extends Report
 
 		$instances = $this->rows? $this->rows : $this->decorator->getListingModels();
 		foreach($instances as $page) {
-			$page->parent_id = $page->parent_id ?: 0;
-
-			$this->tree[$page->parent_id][] = new Row($page->id, $page, $this->decorator->getColumns($page));
+			$index = $inst->parent_id ?: 0;
+			$this->tree[$index][] = new Row($page->id, $page, $this->decorator->getColumns($page));
 		}
 
 		return count($this->tree)? $this->tree[0] : array();